requirements.txt
aiolimiter==1.1.0
alembic==1.13.1
anyio==4.2.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.4
pillow==10.2.0
pycparser==2.21
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.25
tenacity==8.2.3
tornado==6.3.3
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
